requirements.txt
# 核心依赖
numpy>=1.21.0
scipy>=1.7.0
torch>=1.9.0

# 机器人控制
polymetis>=0.1.0
zerorpc>=0.6.0

# 原子操作
atomics>=0.1.0

# 配置和工具
PyYAML>=6.0
click>=8.0.0

# 可选依赖（用于高级功能）
# opencv-python>=4.5.0  # 如果需要图像处理
# matplotlib>=3.5.0     # 如果需要可视化
# tqdm>=4.60.0          # 如果需要进度条
